Reorganize lib/ folder, email-defaults -> email-templates, text-utils -> variables
