+++ app/hooks/use-variable-insertion.ts
@@ -15,8 +15,7 @@ function useVariableInsertion<T extends HTMLInputElement | HTMLTextAreaElement>(
 		}
 	};
 
-	// onBlur matters most here: clicking the dropdown trigger blurs the field
-	// before onInsert runs, so the cursor position must already be captured.
+	// onBlur is the most important
 	const trackingProps = {
 		onSelect: trackCursor,
 		onClick: trackCursor,
@@ -24,10 +23,7 @@ function useVariableInsertion<T extends HTMLInputElement | HTMLTextAreaElement>(
 		onBlur: trackCursor,
 	};
 
-	// Also writes the new value onto the field itself, since uncontrolled
-	// callers (no onChange/value prop) have nothing else applying it. For
-	// controlled callers this is redundant with their own state update, but
-	// harmless — the DOM already matches by the time React re-renders.
+	// Important for uncontrolled inputs (without onChange/value prop)	
 	const insertAtCursor = (currentValue: string, placeholder: string): string => {
 		const { start, end } = cursorRef.current;
 		const text = currentValue.slice(0, start) + placeholder + currentValue.slice(end);
@@ -39,11 +35,10 @@ function useVariableInsertion<T extends HTMLInputElement | HTMLTextAreaElement>(
 		return text;
 	};
 
+	// Specific for DropdownMenu: 
 	// Radix restores focus to the dropdown trigger on close by default, on its
-	// own timing (after any close animation) — later than any rAF/timeout we
-	// could race against. Call this from the dropdown's onCloseAutoFocus
-	// (with the menu's default behavior prevented) instead, so our refocus is
-	// the one that actually sticks.
+	// own timing (after any close animation). Use `restoreFocus` with `onCloseAutoFocus`
+	// (with the menu's default behavior prevented)
 	const restoreFocus = () => {
 		const cursor = pendingCursorRef.current;
 		if (cursor !== null) {
