+++ app/lib/email-templates.ts
@@ -1,8 +1,3 @@
-import {
-  CERTIFICATE_VARIABLE_GROUPS,
-  type VariableGroup,
-} from "./text-utils";
-
 export const EMAIL_TEMPLATES = {
   "notification": {
     label: "Certificate only",
@@ -39,103 +34,6 @@ export type EmailKey = keyof typeof EMAIL_TEMPLATES;
 
 export function isValidEmailKey(key: string | undefined): key is EmailKey {
   return !!key && key in EMAIL_TEMPLATES;
-}
-
-export const EMAIL_KEY_VARIABLE_GROUPS: Record<EmailKey, VariableGroup[]> = {
-  "notification": [
-    ...CERTIFICATE_VARIABLE_GROUPS,
-    {
-      group: "Program & Links",
-      variables: [
-        { placeholder: "program.name", label: "Program name" },
-        { placeholder: "cert.url", label: "Certificate link" },
-      ],
-    },
-  ],
-  "notification-public": [
-    ...CERTIFICATE_VARIABLE_GROUPS,
-    {
-      group: "Program & Links",
-      variables: [
-        { placeholder: "program.name", label: "Program name" },
-        { placeholder: "cert.url", label: "Certificate link" },
-        { placeholder: "cert.loginUrl", label: "Sign in link" },
-        { placeholder: "cert.signAction", label: "Sign action (in/up)" },
-      ],
-    },
-  ],
-  "verify-email": [
-    {
-      group: "User",
-      variables: [
-        { placeholder: "user.firstName", label: "First name" },
-        { placeholder: "user.lastName", label: "Last name" },
-        { placeholder: "user.fullName", label: "Full name" },
-        { placeholder: "verify.url", label: "Verification link" },
-      ],
-    },
-    {
-      group: "Organisation",
-      variables: [
-        { placeholder: "org.name", label: "Organisation name" },
-      ],
-    },
-  ],
-  "password-reset": [
-    {
-      group: "User",
-      variables: [
-        { placeholder: "user.firstName", label: "First name" },
-        { placeholder: "user.lastName", label: "Last name" },
-        { placeholder: "user.fullName", label: "Full name" },
-        { placeholder: "reset.url", label: "Reset link" },
-      ],
-    },
-    {
-      group: "Organisation",
-      variables: [
-        { placeholder: "org.name", label: "Organisation name" },
-      ],
-    },
-  ],
-  "invite": [
-    {
-      group: "Invite",
-      variables: [
-        { placeholder: "invite.firstName", label: "First name" },
-        { placeholder: "invite.lastName", label: "Last name" },
-        { placeholder: "invite.fullName", label: "Full name" },
-        { placeholder: "invite.senderName", label: "Sender name" },
-        { placeholder: "invite.acceptUrl", label: "Accept link" },
-      ],
-    },
-    {
-      group: "Organisation",
-      variables: [
-        { placeholder: "org.name", label: "Organisation name" },
-      ],
-    },
-  ],
-};
-
-// Flags `{variable}` placeholders that aren't in EMAIL_KEY_VARIABLE_GROUPS for this template key.
-export function checkUnknownVariables(key: EmailKey, texts: string[]): string[] {
-  const known = new Set(
-    EMAIL_KEY_VARIABLE_GROUPS[key].flatMap((g) =>
-      g.variables.map((v) => `{${v.placeholder}}`),
-    ),
-  );
-  const unknown = new Set<string>();
-
-  for (const text of texts) {
-    for (const [placeholder] of text.matchAll(/\{[\w.]+\}/g)) {
-      if (!known.has(placeholder)) unknown.add(placeholder);
-    }
-  }
-
-  return [...unknown].map(
-    (placeholder) => `\`${placeholder}\` is not an available variable for this email`,
-  );
 }
 
 export const EMAIL_DEFAULTS: Record<
