+++ app/routes/org.emails.$key.send-preview.ts
@@ -2,7 +2,7 @@ import type { Route } from "./+types/org.emails.$key.send-preview";
 import { redirect } from "react-router";
 
 import { requireSuperAdmin } from "~/lib/auth.server";
-import { isValidEmailKey } from "~/lib/email-defaults";
+import { isValidEmailKey } from "~/lib/email-templates";
 import { sendEmailTemplatePreview } from "~/lib/email.server";
 import { prisma } from "~/lib/prisma.server";
