+++ app/routes/org.emails.$key_.edit.tsx
@@ -4,7 +4,7 @@ import { data, redirect, useActionData } from "react-router";
 import { EmailForm } from "~/components/email-form";
 
 import { requireSuperAdmin } from "~/lib/auth.server";
-import { isValidEmailKey } from "~/lib/email-defaults";
+import { isValidEmailKey } from "~/lib/email-templates";
 import { loadEmailTemplateEditor, saveEmailTemplate } from "~/lib/email.server";
 
 const basePath = "/org/emails";
