+++ app/routes/org.emails._index.tsx
@@ -2,7 +2,7 @@ import type { Route } from "./+types/org.emails._index";
 import { redirect } from "react-router";
 
 import { requireSuperAdmin } from "~/lib/auth.server";
-import { EMAIL_TEMPLATES } from "~/lib/email-defaults";
+import { EMAIL_TEMPLATES } from "~/lib/email-templates";
 
 export async function loader({ request }: Route.LoaderArgs) {
   await requireSuperAdmin(request);
