+++ app/routes/org.program.$programId.emails._index.tsx
@@ -2,7 +2,7 @@ import type { Route } from "./+types/org.program.$programId.emails._index";
 import { redirect } from "react-router";
 
 import { requireAdminWithProgram } from "~/lib/auth.server";
-import { EMAIL_TEMPLATES } from "~/lib/email-defaults";
+import { EMAIL_TEMPLATES } from "~/lib/email-templates";
 
 export async function loader({ request, params }: Route.LoaderArgs) {
   const programId = Number(params.programId);
